refactor(minimald,minvmd): extract shared json_file_layer into mlog crate
The JSON-lines file-log layer (R4.1/R4.2) was defined twice — in
minimald's DaemonLogger and minvmd's init_tracing — with byte-identical
bodies bar a hardcoded vs. parameterized service name. Two copies of a
pinned wire contract drift silently, and the shape test guarded only the
minvmd copy.

Hoist it into a new leaf crate `mlog` both daemons depend on: one
json_file_layer(writer, service_name), one json_log_shape contract test.
The shared version::LONG_VERSION is read internally (identical for every
binary), so only the service name is a parameter. json-subscriber and
opentelemetry-semantic-conventions move off both daemons and become
transitive via mlog.
