+++ crates/minimald/Cargo.toml
@@ -50,8 +50,7 @@ tokio.workspace = true
 tokio-util.workspace = true
 tracing.workspace = true
 tracing-appender.workspace = true
-json-subscriber.workspace = true
-opentelemetry-semantic-conventions.workspace = true
+mlog.workspace = true
 tracing-subscriber.workspace = true
 
 # WireGuard mesh peer (Unit 4, R4.1/R4.7): compiled only under `networking-wg`.
