+++ crates/mlog/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "mlog"
+version = "0.0.1"
+rust-version = "1.91"
+license.workspace = true
+edition.workspace = true
+publish.workspace = true
+
+[dependencies]
+json-subscriber.workspace = true
+opentelemetry-semantic-conventions.workspace = true
+tracing.workspace = true
+tracing-appender.workspace = true
+tracing-subscriber.workspace = true
+version.workspace = true
+
+[dev-dependencies]
+serde_json.workspace = true
